feat(tests/decode_tests): make decode tests more generic
This does not change the test coverage or behavior, it just moves things
around in preparation for reusing the same "decode_tests" architecture
for different sources than the openssl 3.5 files we have now.
